@@ -71,68 +53,75 @@
   "devDependencies": {
     "@commitlint/cli": "^19.6.1",
     "@commitlint/config-conventional": "^19.6.0",
-    "@eslint/compat": "^1.2.4",
-    "@eslint/eslintrc": "^3.2.0",
-    "@eslint/js": "^9.17.0",
+    "@eslint/js": "^9.18.0",
     "@testing-library/dom": "^10.4.0",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.1.0",
     "@testing-library/user-event": "^14.5.2",
     "@types/jest": "^29.5.14",
-    "@types/node": "^22.10.2",
-    "@types/react": "^19.0.2",
+    "@types/node": "^22.10.5",
+    "@types/react": "^19.0.4",
     "@types/react-dom": "^19.0.2",
-    "@typescript-eslint/eslint-plugin": "^8.18.2",
-    "@typescript-eslint/parser": "^8.18.2",
+    "@vitejs/plugin-react-swc": "^3.7.2",
     "commitizen": "^4.3.1",
-    "copy-webpack-plugin": "^12.0.2",
-    "css-loader": "^7.1.2",
-    "css-minimizer-webpack-plugin": "^7.0.0",
     "cz-git": "^1.11.0",
-    "eslint": "^9.17.0",
+    "eslint": "^9.18.0",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-import-resolver-typescript": "^3.7.0",
     "eslint-plugin-import": "^2.31.0",
-    "eslint-plugin-jest": "^28.10.0",
+    "eslint-plugin-prettier": "^5.2.1",
     "eslint-plugin-react": "^7.37.3",
     "eslint-plugin-react-hooks": "^5.1.0",
     "globals": "^15.14.0",
-    "html-webpack-plugin": "^5.6.3",
     "husky": "^9.1.7",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
+    "lint-staged": "^15.3.0",
     "madge": "^8.0.0",
-    "mini-css-extract-plugin": "^2.9.2",
     "miragejs": "^0.1.48",
     "prettier": "^3.4.2",
-    "style-loader": "^4.0.0",
-    "terser-webpack-plugin": "^5.3.11",
     "ts-jest": "^29.2.5",
-    "ts-loader": "^9.5.1",
     "ts-node": "^10.9.2",
     "ts-prune": "^0.10.3",
-    "typescript": "^5.7.2",
-    "webpack": "5.97.1",
-    "webpack-bundle-analyzer": "^4.10.2",
-    "webpack-cli": "^6.0.1",
-    "webpack-dev-server": "^5.2.0",
-    "webpack-merge": "^6.0.1"
+    "typescript": "^5.7.3",
+    "typescript-eslint": "^8.19.1",
+    "vite": "^6.0.7",
+    "vite-plugin-favicons-inject": "^2.2.0",
+    "vite-plugin-static-copy": "^2.2.0"
   },
   "optionalDependencies": {
     "cypress": "^13.17.0",
     "start-server-and-test": "^2.0.9"
   },
-  "engines": {
-    "node": ">=20.0.0",
-    "yarn": ">=1.22.10"
+  "type": "module",
+  "scripts": {
+    "build": "tsc && vite build",
+    "start": "vite --host",
+    "preview": "vite preview",
+    "test": "USE_MOCK_SERVER=true jest --config jest.config.ts",
+    "coverage": "yarn test --coverage --collectCoverageFrom='src/**/*.{tsx,ts}'",
+    "cy:open": "cypress open",
+    "cy:run": "cypress run",
+    "ci": "start-test 3000 cy:run",
+    "lint": "TIMING=1 eslint --cache",
+    "lint-fix": "yarn lint --fix",
+    "format": "prettier --write '*/**/*.{ts,tsx,json,css}'",
+    "check-cycles": "madge --circular --extensions ts,tsx src",
+    "find-unused-modules": "ts-prune",
+    "commit": "git-cz",
+    "tsc:watch": "tsc --watch",
+    "prepare": "husky"
   },
+  "engines": {
+    "node": ">=20.0.0"
+  },
+  "packageManager": "yarn@1.22.19",
   "config": {
     "commitizen": {
       "path": "node_modules/cz-git"
     }
   },
   "browserslist": [
-    " > 0.2%",
+    "> 0.5%",
     "not dead",
     "not op_mini all"
   ]
