@@ -119,4 +119,4 @@ const Configuration = {
   }
 };
 
-module.exports = Configuration;
+export default Configuration;
