@@ -1,11 +1,10 @@
 import { renderHook, render, act } from '@testing-library/react';
 import eventUser from '@testing-library/user-event';
 
-import { useDisplayOptionsState } from '../src/pages/Topology/hooks/useDisplayOptionsState';
-
-import DisplayOptions from '../src/pages/Topology/components/DisplayOptions';
-import { SHOW_DATA_LINKS, SHOW_ROUTER_LINKS } from '../src/pages/Topology/Topology.constants';
 import { Labels } from '../src/config/labels';
+import DisplayOptions from '../src/pages/Topology/components/DisplayOptions';
+import { useDisplayOptionsState } from '../src/pages/Topology/hooks/useDisplayOptionsState';
+import { SHOW_DATA_LINKS, SHOW_ROUTER_LINKS } from '../src/pages/Topology/Topology.constants';
 
 const mockOnSelected = jest.fn();
 const defaultSelected: string[] = [];
