@@ -1,8 +1,8 @@
 import { render, screen } from '@testing-library/react';
 import eventUser from '@testing-library/user-event';
 
-import DisplayResources from '../src/pages/Topology/components/DisplayResources';
 import { Labels } from '../src/config/labels';
+import DisplayResources from '../src/pages/Topology/components/DisplayResources';
 
 describe('DisplayResources', () => {
   const onSelectMock = jest.fn();
