@@ -4,25 +4,26 @@ import { Button } from '@patternfly/react-core';
 import { render, screen, waitForElementToBeRemoved } from '@testing-library/react';
 import eventUser from '@testing-library/user-event';
 import { Server } from 'miragejs';
-import { mockNavigate } from '../jest.mock.router';
+
+import { mockNavigate } from '../jest.setup';
 import processesPairsData from '../mocks/data/PROCESS_PAIRS.json';
 import processesData from '../mocks/data/PROCESSES.json';
 import servicesData from '../mocks/data/SERVICES.json';
 import { loadMockServer } from '../mocks/server';
+import { extendedProcessResponse } from '../mocks/server.API';
 import { waitForElementToBeRemovedTimeout } from '../src/config/app';
+import { Labels } from '../src/config/labels';
 import { getTestsIds } from '../src/config/testIds';
 import LoadingPage from '../src/core/components/SkLoading';
-import { Providers } from '../src/providers';
 import { ProcessesRoutesPaths } from '../src/pages/Processes/Processes.enum';
 import { TopoloyDetailsProps } from '../src/pages/Topology/components/TopologyDetails';
 import TopologyProcesses from '../src/pages/Topology/components/TopologyProcesses';
 import * as useTopologyState from '../src/pages/Topology/hooks/useTopologyState';
 import { TopologyController } from '../src/pages/Topology/services';
 import { convertProcessToNode } from '../src/pages/Topology/services/topologyProcessController';
+import { Providers } from '../src/providers';
 import { SkGraphProps } from '../src/types/Graph.interfaces';
 import { ProcessPairsResponse, ServiceResponse } from '../src/types/REST.interfaces';
-import { Labels } from '../src/config/labels';
-import { extendedProcessResponse } from '../mocks/server.API';
 
 const processesResults = processesData.results as extendedProcessResponse[];
 const processesPairsResults = processesPairsData.results as ProcessPairsResponse[];
