@@ -11,11 +11,11 @@ import { loadMockServer } from '../mocks/server';
 import { waitForElementToBeRemovedTimeout } from '../src/config/app';
 import { getTestsIds } from '../src/config/testIds';
 import LoadingPage from '../src/core/components/SkLoading';
-import { Providers } from '../src/providers';
 import TopologySite from '../src/pages/Topology/components/TopologySite';
 import { TopologyController } from '../src/pages/Topology/services';
 import { convertSiteToNode } from '../src/pages/Topology/services/topologySiteController';
 import { SHOW_DATA_LINKS } from '../src/pages/Topology/Topology.constants';
+import { Providers } from '../src/providers';
 import { SkGraphProps } from '../src/types/Graph.interfaces';
 import { PairsResponse, SiteResponse } from '../src/types/REST.interfaces';
 
