@@ -40,12 +40,14 @@ const SkChartPie: FC<SkChartPieProps> = function ({
   }
 
   useEffect(() => {
-    if (chartContainerRef.current) {
-      observer.current = getResizeObserver(chartContainerRef.current, handleResize);
-      handleResize();
-
-      () => observer.current();
+    if (!chartContainerRef.current) {
+      return;
     }
+
+    observer.current = getResizeObserver(chartContainerRef.current, handleResize);
+    handleResize();
+
+    return () => observer.current();
   }, []);
 
   const total = data.reduce((acc, { y }) => acc + y, 0);
