@@ -12,8 +12,13 @@ createRoot(rootElement).render(
   </Providers>
 );
 
-if (!process.env.COLLECTOR_URL && (process.env.NODE_ENV === 'development' || !!process.env.ENABLE_MOCK_SERVER)) {
-  const mockServer = require('../mocks/server');
-
-  mockServer.loadMockServer();
+if (!process.env.OBSERVER_URL && (process.env.NODE_ENV === 'development' || !!process.env.USE_MOCK_SERVER)) {
+  import('../mocks/server')
+    .then((module) => {
+      module.loadMockServer();
+    })
+    .catch((error) => {
+      // eslint-disable-next-line no-console
+      console.error('Error loading mock server:', error);
+    });
 }
